Add a second host.
It is used as a GitHub Actions runner.

It needs to be deployed in two steps: the initial deployment step, where
no secrets are deployed, then a second step that configure the runner.
This requires a secretm which is encrypted using the machine SSH host
key (which is available only after the initial deployment).