Duplicate the common code.
And try to use it in two deployments: common-1, and common-2.

They both use Terragrunt. The later one is using it together with a
main.tf file, the first one is only using a terragrunt.hcl file.

The second one can be applied using `terragrunt apply`, but not the
first one: it seems it can't pass correctly the `default.nix` file to
the build-nix module.